+++ spec/rspec/core/configuration_options_spec.rb
@@ -102,13 +102,12 @@ RSpec.describe RSpec::Core::ConfigurationOptions, :isolated_directory => true, :
       opts.configure(config)
     end
 
-    it "sets pattern before loading specs" do
-      opts = config_options_object(*%w[--pattern **/*.spec])
+    it "sets pattern before `requires` so users can check `files_to_run` in a `spec_helper` loaded by `--require`" do
+      opts = config_options_object(*%w[--require spec_helpe --pattern **/*.spec])
       config = RSpec::Core::Configuration.new
       expect(config).to receive(:force).with(:pattern => '**/*.spec').ordered
-      expect(config).to receive(:get_files_to_run).ordered
+      expect(config).to receive(:requires=).ordered
       opts.configure(config)
-      config.files_to_run
     end
 
     it "assigns inclusion_filter" do